requirements.txt
Django==1.6.3
django-allauth==0.16.1
django-autoslug==1.7.2
django-braces==1.4.0
django-crispy-forms==1.4.0
oauthlib==0.6.1
python-openid==2.2.5
requests==2.2.1
requests-oauthlib==0.4.0
six==1.6.1
wsgiref==0.1.2
